+++ test/td_dd/loader/metadata_test.exs
@@ -6,6 +6,7 @@ defmodule TdDd.Loader.MetadataTest do
   alias TdDd.CSV.Reader
   alias TdDd.DataStructures.StructureMetadata
   alias TdDd.Loader
+  alias TdDd.Loader.Metadata
   alias TdDd.Repo
 
   @structure_import_schema Application.compile_env(:td_dd, :metadata)[:structure_import_schema]
@@ -15,7 +16,7 @@ defmodule TdDd.Loader.MetadataTest do
   @relation_import_schema Application.compile_env(:td_dd, :metadata)[:relation_import_schema]
   @relation_import_required Application.compile_env(:td_dd, :metadata)[:relation_import_required]
 
-  describe "load/2" do
+  describe "Loader.load/2" do
     test "inserts and logically deletes mutable metadata" do
       %{id: system_id} = insert(:system)
 
@@ -82,6 +83,29 @@ defmodule TdDd.Loader.MetadataTest do
                  version: 1
                }
              ]
+    end
+  end
+
+  describe "missing_external_ids/4" do
+    test "identifies external_ids which don't exist in the system" do
+      %{external_id: external_id1, system: system} = insert(:data_structure)
+      %{external_id: external_id2} = insert(:data_structure)
+
+      records = [
+        %{external_id: "foo"},
+        %{external_id: external_id1},
+        %{external_id: external_id2}
+      ]
+
+      assert {:error, ["foo", ^external_id2]} =
+               Metadata.missing_external_ids(Repo, %{}, records, system)
+    end
+
+    test "returns ok and empty list if all external_ids exist" do
+      %{external_id: external_id1, system: system} = insert(:data_structure)
+
+      records = [%{external_id: external_id1}]
+      assert {:ok, []} = Metadata.missing_external_ids(Repo, %{}, records, system)
     end
   end