+++ test/td_dd_web/controllers/system_metadata_controller_test.exs
@@ -71,6 +71,28 @@ defmodule TdDdWeb.SystemMetadataControllerTest do
     end
   end
 
+  describe "PATCH /api/systems/:external_id/metadata" do
+    @tag authentication: [role: "service"]
+    test "calls worker with valid metadata json data", %{conn: conn} do
+      %{id: system_id, external_id: system_external_id} = insert(:system)
+
+      body = %{"op" => "replace", "values" => [%{"foo" => "bar"}]}
+
+      expect(TdDd.Loader.MockWorker, :load, fn %{id: ^system_id},
+                                               %{"system_id" => _} = params,
+                                               audit,
+                                               _opts ->
+        assert body == Map.delete(params, "system_id")
+        assert %{ts: _, last_change_by: _} = audit
+        :ok
+      end)
+
+      assert conn
+             |> patch(Routes.system_metadata_path(conn, :update, system_external_id, body))
+             |> response(:accepted)
+    end
+  end
+
   defp upload(path) do
     %Plug.Upload{path: path, filename: Path.basename(path), content_type: "text/csv"}
   end